adding hover css on icon
