+++ package.json
@@ -13,6 +13,7 @@
     "bootstrap": "^5.3.3",
     "react": "^18.2.0",
     "react-bootstrap": "^2.10.2",
+    "react-bootstrap-icons": "^1.11.4",
     "react-dom": "^18.2.0"
   },
   "devDependencies": {
