+++ src/components/auctionCreationContainer/AuctionSelectionDropdowns.css
@@ -24,7 +24,7 @@
   padding: 10px 15px;
   border-radius: 10px;
   margin-top: 2rem;
-  width: 19rem;
+  width: 15rem;
   font-weight: bold;
   text-align: center;
 }
