--- src/components/auctionCreationContainer/InfoFieldsCreateAuction.jsx
@@ -1,9 +1,0 @@
-
-
-const InfoFieldsCreateAuction = () => {
-  return (
-    <div>InfoFieldsCreateAuction</div>
-  )
-}
-
-export default InfoFieldsCreateAuction;
