--- src/components/auctionCreationContainer/RandomPic.css
@@ -1,22 +1,0 @@
-.form {
-  width: 12rem;
-  height: 2.5rem;
-  background-color: aliceblue;
-  border: none;
-  border-radius: 5px;
-  display: flex;
-  align-items: center;
-  justify-content: flex-end;
-  padding-right: 1rem;
-  margin-top: 2rem;
-  margin-left: 2rem;
-}
-
-button {
-  display: flex;
-  height: 2rem;
-  width: 2rem;
-  border: 1px solid black;
-  cursor: pointer;
-  background-color: white;
-}
