--- src/components/auctionCreationContainer/RandomPic.jsx
@@ -1,14 +1,0 @@
-import "../auctionCreationContainer/RandomPic.css";
-
-const RandomPic = () => {
-  return (
-    <form className="form">
-      <button type="button"></button>
-      <div className="random">
-        <label>Random Picture</label>
-      </div>
-    </form>
-  );
-};
-
-export default RandomPic;
